# docs/openapi/backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transit Delay Predictor — Backend",
    "description": "Оркестрация: приём телеметрии NDTP, прогнозы ML-сервиса, инциденты, API и WebSocket для дашборда.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/telemetry": {
      "post": {
        "summary": "Приём батча телеметрии от NDTP-сервера",
        "operationId": "telemetry_api_v1_telemetry_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/NavRecordIn"
                },
                "type": "array",
                "title": "Records"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/whatif": {
      "post": {
        "summary": "What-if: как изменится прогноз при действии диспетчера",
        "operationId": "whatif_api_v1_whatif_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhatIfIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/segments": {
      "get": {
        "summary": "Проблемные участки: где ТС сегодня теряют время (по фактическим прибытиям)",
        "operationId": "segments_api_v1_segments_get",
        "parameters": [
          {
            "name": "min_obs",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 2,
              "title": "Min Obs"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "NavRecordIn": {
        "properties": {
          "unit_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit Id",
            "description": "ID бортового терминала (peerAddress)"
          },
          "tr_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tr Id",
            "description": "ID ТС, если известен"
          },
          "t": {
            "type": "number",
            "title": "T",
            "description": "Unix-время фиксации, с"
          },
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "default": true
          },
          "lon": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          }
        },
        "type": "object",
        "required": [
          "t"
        ],
        "title": "NavRecordIn",
        "description": "Навигационная запись (ячейка NDTP ``G6CellNav00`` после парсинга)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WhatIfIn": {
        "properties": {
          "tr_id": {
            "type": "integer",
            "title": "Tr Id"
          },
          "delay_shift_s": {
            "type": "number",
            "title": "Delay Shift S",
            "description": "ТС опоздает ещё на столько секунд (+) / нагонит (−)",
            "default": 0
          },
          "extra_layover_min": {
            "type": "number",
            "title": "Extra Layover Min",
            "description": "добавить минут отстоя на конечной (если отстой есть на пути)",
            "default": 0
          },
          "reserve_vehicle": {
            "type": "boolean",
            "title": "Reserve Vehicle",
            "description": "выпустить резервное ТС по графику",
            "default": false
          },
          "overrides": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Overrides",
            "description": "прямое переопределение признаков"
          }
        },
        "type": "object",
        "required": [
          "tr_id"
        ],
        "title": "WhatIfIn",
        "description": "Сценарий диспетчера для ТС (к его последнему прогнозу)."
      }
    }
  }
}